Add files via upload
Body mass index (BMI) is a measure of body fat based on height as well as the weight. It had been applies to adult men and women. It will ask you two quesions to you. For example :- Enter your weight :- Enter your height :- After that it will calculate the BMI and tell weather you are fit or not. At the same time you will asked to save it in file or not